move reader to __init__.py, correct some tests
