+++ act/test/test_act_utils.py
@@ -1,7 +1,7 @@
 import unittest
 
 from nari.ext.act.actlogutils import ActLogChecksumType, validate_checksum
-from nari.ext.act.actlogutils.exceptions import InvalidActChecksumAlgorithm
+from nari.ext.act.exceptions import InvalidActChecksumAlgorithm
 
 
 class TestLineChecksum(unittest.TestCase):
