+++ act/test/test_marker_parsing.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from nari.types.event.markers import PlayerMarker, MarkerOperation, OverheadMarker
 from nari.ext.act.actlogutils.targetmarker import targetmarker_from_logline
-from nari.ext.act.actlogutils.exceptions import InvalidMarkerID
+from nari.ext.act.exceptions import InvalidMarkerID
 
 
 class TestTargetMarkerFromACTLogLine(unittest.TestCase):
